Separate unit tests into dedicated test file
- Move unit tests from src/unfold.rs to tests/unit_test.rs
- Make split_by_coloncolon and unfold_curly_bracket public for testing
- Update README to reflect new project structure
- Better separation of concerns: application code vs test code
